+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### WIP
 
+* More SRTP crypto suites (PR #837).
 * Improve `EnhancedEventEmitter` (PR #836).
 * `WebRtcServer`: A new class that brings to `WebRtcTransports` the ability to listen on a single UDP/TCP port (PR #834).
 * `TransportCongestionControlClient`: Allow setting max outgoing bitrate before created (PR #833).